pyproject.toml
[project]
name = "python-project-50"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.main:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
